Update runPrimeSearch.py
Several tests have been run to check the efficiency of the diffent isPrimeSearch algorithms.

In all the intervals checked, CheckForPrime6 is the fastest algorithm. Both from an overall perspective and on a per type of loop perspective.

Therefore, this should be used as the test in the RunPrimeSearch progam.

This commit changes the search to checkForPrime6